Test: di extension のテストヘルパーを共有 testkit へ集約
6 パッケージで重複していた requireProvidesOne を
internal/di/server/extension/testkit.RequireProvidesOne に集約し、各
module_graph_test.go から重複定義を除去。
